requirements.txt
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.1
moviepy==1.0.3
scikit-image==0.21.0
matplotlib==3.7.2
tqdm==4.66.1
vidgear==0.3.0
python-telegram-bot==20.7
python-dotenv==1.0.0
yadisk==1.4.0
aiofiles==23.2.1
asyncio
concurrent.futures
pathlib
datetime
uuid
# Social media stats libraries
instaloader==4.14.2
TikTokApi==7.2.1
vk-api==11.9.9
requests==2.32.5
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
